requirements.txt
# requirements.txt

# Deep learning — NLP model (DistilBERT)
torch==2.1.0
transformers==4.36.0
accelerate==0.25.0
datasets==2.16.0

# ML — LightGBM classifier
lightgbm==4.1.0
scikit-learn==1.3.2
numpy==1.24.0
pandas==2.1.0

# API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
plyer==2.1.0

# Database
sqlalchemy==2.0.23

# Email parsing — imaplib is stdlib, no pip install needed
beautifulsoup4==4.12.2

# Domain intelligence
tldextract==5.1.1
requests==2.31.0
python-whois==0.9.4

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2